+++ CHANGELOG.md
@@ -17,6 +17,10 @@
 - `Drivers`:
     - SPI can create RX and TX DMA Streams.
     - Adjusts PLL defines for L0 devices.
+- `Graphics`:
+    - Adds << operator to display canvas for strings.
+
+#### [&rarrb; v3.4.0](https://github.com/SebastianOberschwendtner/OTOS/releases/tag/v3.4.0)
 
 ### Fixed Issues:
 - #12: Update core_stm32 for STM32L0 Devices